requirements.txt
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.46.0
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
protobuf==6.33.2
PyMatting==1.1.14
pyreadline3==3.5.4
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.16.3
sympy==1.14.0
tifffile==2025.12.20
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
gunicorn==23.0.0